requirements.txt
sqlalchemy==2.0.40
loguru==0.7.3
pydantic==2.11.3
psycopg2-binary==2.9.10
fastapi==0.115.12
python-dotenv==0.20.0
uvicorn==0.34.2

